pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "community>=1.0.0b1",
    "fastapi>=0.115.12",
    "fpdf>=1.7.2",
    "google-generativeai>=0.8.4",
    "matplotlib>=3.10.1",
    "neo4j>=5.28.1",
    "networkx>=3.4.2",
    "nltk>=3.9.1",
    "node2vec>=0.4.3",
    "numpy>=2.2.4",
    "openai>=1.72.0",
    "pandas>=2.2.3",
    "passlib[bcrypt]>=1.7.4",
    "pdfkit>=1.0.0",
    "plotly>=6.0.1",
    "psycopg2-binary>=2.9.10",
    "py2neo>=2021.2.4",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.4.0",
    "python-multipart>=0.0.20",
    "reportlab>=4.3.1",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.2",
    "seaborn>=0.13.2",
    "streamlit>=1.44.1",
    "trafilatura>=2.0.0",
    "uvicorn>=0.34.0",
    "weasyprint>=65.0",
]
